% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{get_species}
\alias{get_species}
\title{The species that is set}
\usage{
get_species()
}
\value{
A list (\code{name}, \code{organism}, \code{genome}, \code{sizes}) or \code{NULL} if no
species was set with \code{\link[=set_species]{set_species()}}.
}
\description{
The species that is set
}
\examples{
get_species()
}
