% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_zoom_connector}
\alias{plot_zoom_connector}
\title{Connector lines from the genome-wide plot to the regional panels}
\usage{
plot_zoom_connector(manhattan, region, colour = "grey45")
}
\arguments{
\item{manhattan}{A plot from \code{\link[=plot_manhattan]{plot_manhattan()}}.}

\item{region}{The region shown in the panels below.}

\item{colour}{Line colour.}
}
\value{
A ggplot object to place between the Manhattan plot and the
regional panels (see \code{\link[=combine_panels]{combine_panels()}}).
}
\description{
Connector lines from the genome-wide plot to the regional panels
}
