% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{write_blocks}
\alias{write_blocks}
\title{Write LD blocks in PLINK \code{.blocks.det} format}
\usage{
write_blocks(blocks, file)
}
\arguments{
\item{blocks}{Output of \code{\link[=find_ld_blocks]{find_ld_blocks()}} or \code{\link[=read_blocks]{read_blocks()}}.}

\item{file}{Output path.}
}
\value{
The file path (invisibly).
}
\description{
Write LD blocks in PLINK \code{.blocks.det} format
}
